Markdown
Chore: Add solargraph and update dependencies

This commit introduces the solargraph gem as a development dependency. Solargraph provides static analysis and language server features for Ruby.

The addition of solargraph has also resulted in an update to the Bundler version in the Gemfile.lock and several updates to other gem dependencies.